requirements.txt
requests==2.31.0
urllib3==2.1.0
pandas==2.2.2
matplotlib==3.8.4
numpy==1.26.4
seaborn==0.12.2
plotly==5.21.0
scipy==1.11.4
openpyxl==3.1.2
Pillow==10.2.0
python-dateutil==2.8.2
PyYAML=6.0.1
html2image==2.0.7
jinja2==3.1.6
flask=3.1.1